+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Changelog
+ # Changelog
 
 All notable changes to this project will be documented in this file.
 
@@ -10,13 +10,23 @@ how a consumer would use the library (e.g. adding unit tests, updating documenta
 
 ## Unreleased
 
+### Fixed
+
+- Issue where `code42 profile delete` was allowed even deleting the default profile is not allowed.
+
 ### Added
+
+- `code42 departing-employee list` command.
+
+- `code42 high-risk-employee list` command.
 
 - `code42 audit-logs` commands:
     - `search` to search for audit-logs.
     - `send-to` to send audit-logs to server.
 
 ### Changed
+
+- `profile_name` argument is now required for `code42 profile delete`, as it was meant to be.
 
 - The `--advanced-query` option on `alerts search` and `security-data (search|send-to)` commands has been updated:
     - It can now accept the query as a JSON string or as the path to a file containing the JSON query.