+++ CHANGELOG.md
@@ -12,7 +12,7 @@ how a consumer would use the library (e.g. adding unit tests, updating documenta
 
 ### Changed
 
-- The error text when removing an employee from a detection list now references the employee 
+- The error text when removing an employee from a detection list now references the employee
   by ID rather the username.
 
 ### Fixed